services/llm-api/pyproject.toml
[project]
name = "lixpi-llm-api"
version = "0.1.0"
description = "Lixpi LLM API Service - Python-based microservice for AI model interactions"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "langgraph>=0.2.45",
    "langchain>=0.3.7",
    "langchain-openai>=0.2.9",
    "langchain-anthropic>=0.2.4",
    "nats-py>=2.9.0",
    "nkeys>=0.1.0",
    "openai>=1.54.0",
    "anthropic>=0.39.0",
    "pyjwt>=2.9.0",
    "python-jose[cryptography]>=3.3.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "lixpi-constants @ file:///app/packages/lixpi/constants/python",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]
